requirements.txt
pkg_resources==0.0.0
simple-websocket-server==0.4.2
websocket-client==1.5.1
